Separate credentials dialog for certificate and key
